Думаю над авторзацией

+++ back_end/src/main/java/com/example/back_end/service/impl/CourseServiceImpl.java
@@ -1,6 +1,6 @@
 package com.example.back_end.service.impl;
 
-import com.example.back_end.exeptiom.RestBadRequestException;
+import com.example.back_end.exeption.RestBadRequestException;
 import com.example.back_end.persistence.entity.Course;
 import com.example.back_end.persistence.repository.CourseRepository;
 import com.example.back_end.service.CourseService;
@@ -11,6 +11,7 @@ import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
 @Service
 public class CourseServiceImpl implements CourseService {
     private CourseRepository repository;
@@ -28,25 +29,31 @@ public class CourseServiceImpl implements CourseService {
     @Override
     @Transactional(isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRES_NEW, rollbackFor = Exception.class)
     public void create(Course course) {
-        repository.save(course);
+        if (!repository.existsById(course.getId())) {
+            repository.save(course);
+        } else throw new RestBadRequestException("This course is already exist");
     }
 
     @Transactional(isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRES_NEW, rollbackFor = Exception.class)
     @Override
     public void update(Course course) {
-        repository.save(course);
+        if (repository.existsById(course.getId())) {
+            repository.save(course);
+        } else throw new RestBadRequestException("This course is already exist");
     }
 
     @Transactional(isolation = Isolation.READ_COMMITTED, propagation = Propagation.REQUIRES_NEW, rollbackFor = Exception.class)
     @Override
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+        } else throw new RestBadRequestException("This course is not exist");
     }
 
     @Override
     @Transactional(readOnly = true)
     public Course findById(Long id) {
-        return repository.findById(id).orElseThrow(() -> new RestBadRequestException("This course not exist"));
+        return repository.findById(id).orElseThrow(() -> new RestBadRequestException("This course is not exist"));
     }
 
     @Override
